add test coverage to package.json
